pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
openai = "^0.27.4"
telegram = "^0.0.1"
pydub = "^0.25.1"
PyOgg = "^0.6.14-alpha.1"
cloudconvert = "^2.1.0"
python-dotenv = "^1.0.0"
imageio = "^2.27.0"
imageio-ffmpeg = "^0.4.8"
audioread = "^3.0.0"
lameenc = "^1.4.2"
av = "^10.0.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"